+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed Header Navigation in Small Screens
 - Fixed Blog search
+- Fixed Header Link Styling
 
 ### Changed